alarms_service/requirements.txt
Flask
paho-mqtt
SQLAlchemy
psycopg2-binary
flask_sqlalchemy
python-dotenv